+++ docs/fr_FR/Changelog.rst
@@ -1,6 +1,9 @@
 ChangeLog
 =========
 
+- Interne: Parser: Ajout infos debug pour support Xiaomi.
+- Modeles: Mise-à-jour 'sensor_cube' + 'sensor_cube.aqgl01'.
+
 230126-BETA-2
 -------------